razorpay_django_python/requirements.txt
django
djangorestframework
markdown
django-filter
gunicorn
pyjwt
dj-database-url
whitenoise
django-cors-headers
psycopg2-binary
requests
Pillow
razorpay